+++ app/src/main/java/com/moutamid/misscaddie/adapters/RequestesAdapter.java
@@ -54,7 +54,7 @@ public class RequestesAdapter extends RecyclerView.Adapter<RequestesAdapter.VH> 
     public void onBindViewHolder(@NonNull VH holder, int position) {
         RequestsModel model = itemList.get(position);
         String serviceList = "";
-        int price = 0;
+        double price = 0.0;
 
       //  holder.name.setText(model.getName());
         holder.address.setText(model.getAddress());
@@ -77,7 +77,7 @@ public class RequestesAdapter extends RecyclerView.Adapter<RequestesAdapter.VH> 
         for (int i=0; i < model.getTableRows().size(); i++){
             String service = model.getTableRows().get(i).getTitle() +
                     " (USD$" + model.getTableRows().get(i).getPrice() + ")";
-            price = price + Integer.parseInt(model.getTableRows().get(i).getPrice());
+            price = price + Double.parseDouble(model.getTableRows().get(i).getPrice());
             if (i==2){
                 serviceList = serviceList + "\t\t" + "more";
                 break;
